+++ app/src/main/java/com/mutualmobile/praxis/datasource/JokesRemoteSource.kt
@@ -1,9 +1,10 @@
-package com.mutualmobile.praxis.data.sources
+package com.mutualmobile.praxis.datasource
 
 import com.mutualmobile.praxis.data.SafeResult
 import com.mutualmobile.praxis.data.remote.JokeApiService
 import com.mutualmobile.praxis.domain.responsemodels.JokeListResponse
 import com.mutualmobile.praxis.data.remote.safeApiCall
+import com.mutualmobile.praxis.data.sources.IJokesRemoteSource
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
