+++ app/src/main/java/com/mutualmobile/praxis/injection/component/AppComponent.kt
@@ -3,7 +3,7 @@ package com.mutualmobile.praxis.injection.component
 import android.content.Context
 import com.mutualmobile.praxis.BaseApplication
 import com.mutualmobile.praxis.data.injection.RepositoryModule
-import com.mutualmobile.praxis.data.injection.SourcesModule
+import com.mutualmobile.praxis.injection.module.SourcesModule
 import com.mutualmobile.praxis.data.injection.UseCaseModule
 import com.mutualmobile.praxis.injection.module.ActivityBindingModule
 import com.mutualmobile.praxis.injection.module.AppModule
